refactor(json): route applyPatch helpers through the escape entry point
`assertSafeFromPointer` and `ensureOwnedFromPath` reached `unescapePath` at
`escape/unescapePath.ts` — a sibling fractal's internal file. `escape/index.ts`
exports the symbol by name, so the crossing had no reason to exist
(filid_fractal-boundaries §3).

Relative specifiers are used rather than the `@/json/*` alias so the adapter can
resolve the edge; alias references are dropped from the dependency graph.

Cross-review raised this as part of STR-030, which also attributed the fractal's
`indeterminate` dependency-graph certainty to these alias imports. That
attribution is wrong and this commit does not fix it — see the resolve record.

Verified: `yarn workspace @winglet/json typecheck` and
`yarn workspace @canard/schema-form typecheck` pass; json tests 551/551.
